pyproject.toml
[project]
name = "brainless-embedding-service"
version = "0.1.0"
description = ""
authors = [{ name = "Eugene Cheltsov", email = "chill.icp@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi[all]>=0.121.3",
    "pydantic-settings>=2.12.0",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]

[dependency-groups]
dev = ["mypy>=1.18.2", "ruff>=0.14.6", "vulture>=2.14"]
